move test to integration test
